dfns naming

+++ src/stores/DfnsStore.ts
@@ -1,6 +1,7 @@
 import { CredentialInfo } from "@dfns/sdk/codegen/datamodel/Auth";
-import { Wallet } from "@dfns/sdk/codegen/datamodel/Wallets";
+import { Wallet, WalletAsset } from "@dfns/sdk/codegen/datamodel/Wallets";
 import { createStore } from "@stencil/store";
+
 import LocalStorageService, {
 	DFNS_ACTIVE_WALLET,
 	DFNS_CREDENTIALS,
@@ -8,6 +9,10 @@ import LocalStorageService, {
 	OAUTH_ACCESS_TOKEN,
 } from "../services/LocalStorageService";
 import { EventEmitter } from "../services/EventEmitter";
+import IWalletInterface from "../services/wallet/IWalletInterface";
+import { BlockchainNetwork } from "@dfns/sdk/codegen/datamodel/Wallets";
+import { IColors } from "../common/interfaces/IColors";
+import { EThemeModeType } from "../common/enums/themes-enums";
 
 interface DfnsState {
 	appName: string | null;
@@ -15,11 +20,32 @@ interface DfnsState {
 	dfnsHost: string | null;
 	appId: string | null;
 	rpId: string | null;
+	network: BlockchainNetwork | null;
 	dfnsUserToken: string | null;
 	wallet: Wallet | null;
 	oauthAccessToken: string | null;
 	appLogoUrl: string | null;
 	credentials: CredentialInfo[];
+	defaultDevice: "mobile" | "desktop" | null;
+	assets: WalletAsset[];
+	googleEnabled: boolean;
+	googleClientId: string;
+	customButtonEnabled: boolean;
+	customButtonText: string;
+	customButtonIcon: string | null;
+	customButtonCallback: () => any | null;
+	primaryColor: string | null;
+	theme: EThemeModeType;
+	lang: string;
+	walletService: IWalletInterface | null;
+	walletConnectEnabled: boolean;
+	walletConnectProjectId: string;
+	colors: IColors;
+	showWalletValidation: boolean;
+	disableLogoutUI: boolean;
+	activateRecovery: boolean;
+	showRecoverySetupAtWalletCreation: boolean;
+	showRecoverySetupAfterRecoverAccount: boolean;
 }
 
 const dfnsStoreEvent = new EventEmitter<DfnsState>();
@@ -35,48 +61,72 @@ const { state } = createStore<DfnsState>({
 	appLogoUrl: null,
 	oauthAccessToken: null,
 	credentials: [],
+	assets: [],
+	googleClientId: "",
+	googleEnabled: false,
+	customButtonEnabled: false,
+	customButtonText: "",
+	customButtonIcon: null,
+	customButtonCallback: null,
+	primaryColor: null,
+	theme: EThemeModeType.LIGHT,
+	lang: "en",
+	walletService: null,
+	network: null,
+	walletConnectEnabled: false,
+	walletConnectProjectId: "",
+	defaultDevice: null,
+	colors: null,
+	showWalletValidation: false,
+	disableLogoutUI: false,
+	activateRecovery: false,
+	showRecoverySetupAtWalletCreation: false,
+	showRecoverySetupAfterRecoverAccount: false,
 });
 
 function setValue<T extends keyof DfnsState>(key: T, value: DfnsState[T]) {
-	if (key === "dfnsUserToken") {
-		LocalStorageService.getInstance().items[DFNS_END_USER_TOKEN].set(value as string);
-	}
-	if (key === "oauthAccessToken") {
-		LocalStorageService.getInstance().items[OAUTH_ACCESS_TOKEN].set(value as string);
-	}
-	if (key === "wallet") {
-		LocalStorageService.getInstance().items[DFNS_ACTIVE_WALLET].set(value as Wallet);
-	}
-	if (key === "credentials") {
-		LocalStorageService.getInstance().items[DFNS_CREDENTIALS].set(value as CredentialInfo[]);
+	switch (key) {
+		case "dfnsUserToken":
+			if (!value) {
+				LocalStorageService.getInstance().items[DFNS_END_USER_TOKEN].delete();
+				break;
+			}
+			LocalStorageService.getInstance().items[DFNS_END_USER_TOKEN].set(value as string);
+			break;
+		case "oauthAccessToken":
+			if (!value) {
+				LocalStorageService.getInstance().items[OAUTH_ACCESS_TOKEN].delete();
+				break;
+			}
+			LocalStorageService.getInstance().items[OAUTH_ACCESS_TOKEN].set(value as string);
+			break;
+		case "wallet":
+			if (!value) {
+				LocalStorageService.getInstance().items[DFNS_ACTIVE_WALLET].delete();
+				break;
+			}
+			LocalStorageService.getInstance().items[DFNS_ACTIVE_WALLET].set(value as Wallet);
+			break;
+		case "credentials":
+			if (!value) {
+				LocalStorageService.getInstance().items[DFNS_CREDENTIALS].delete();
+				break;
+			}
+			LocalStorageService.getInstance().items[DFNS_CREDENTIALS].set(value as CredentialInfo[]);
+			break;
 	}
 	state[key] = value;
 	dfnsStoreEvent.emit("changed", state);
 }
 
-function disconnect() {
-	state.dfnsUserToken = null;
-	state.oauthAccessToken = null;
-	state.wallet = null;
-	state.credentials = [];
-	LocalStorageService.getInstance().items[DFNS_END_USER_TOKEN].delete();
-	LocalStorageService.getInstance().items[OAUTH_ACCESS_TOKEN].delete();
-	LocalStorageService.getInstance().items[DFNS_ACTIVE_WALLET].delete();
-	LocalStorageService.getInstance().items[DFNS_CREDENTIALS].delete();
-	dfnsStoreEvent.emit("changed", state);
-	dfnsStoreEvent.emit("disconnected", state);
-}
-
 const dfnsStore: {
 	state: Readonly<DfnsState>;
 	setValue: <T extends keyof DfnsState>(key: T, value: DfnsState[T]) => void;
-	disconnect: () => void;
 	dfnsStoreEvent: EventEmitter<DfnsState>;
 } = {
 	state,
 	setValue,
-	disconnect,
-	dfnsStoreEvent
+	dfnsStoreEvent,
 };
 
 export default dfnsStore;
